Isolate react-app package
