+++ cmd/alertmanager/main.go
@@ -57,6 +57,7 @@ import (
 	"github.com/prometheus/alertmanager/timeinterval"
 	"github.com/prometheus/alertmanager/types"
 	"github.com/prometheus/alertmanager/ui"
+	reactApp "github.com/prometheus/alertmanager/ui/react-app"
 )
 
 var (
@@ -495,6 +496,7 @@ func run() int {
 	webReload := make(chan chan error)
 
 	ui.Register(router, webReload, logger)
+	reactApp.Register(router, logger)
 
 	mux := api.Register(router, *routePrefix)
