+++ ui/web.go
@@ -15,7 +15,6 @@ package ui
 
 import (
 	"fmt"
-	"io"
 	"net/http"
 	_ "net/http/pprof" // Comment this line to disable pprof endpoint.
 	"path"
@@ -23,16 +22,9 @@ import (
 	"github.com/go-kit/log"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/prometheus/common/route"
-	"github.com/prometheus/common/server"
 
 	"github.com/prometheus/alertmanager/asset"
-	reactApp "github.com/prometheus/alertmanager/ui/react-app"
 )
-
-var reactRouterPaths = []string{
-	"/",
-	"/status",
-}
 
 // Register registers handlers to serve files for the web interface.
 func Register(r *route.Router, reloadCh chan<- chan error, logger log.Logger) {
@@ -68,37 +60,6 @@ func Register(r *route.Router, reloadCh chan<- chan error, logger log.Logger) {
 		req.URL.Path = path.Join("/static/lib", route.Param(req.Context(), "path"))
 		fs := http.FileServer(asset.Assets)
 		fs.ServeHTTP(w, req)
-	})
-
-	serveReactApp := func(w http.ResponseWriter, r *http.Request) {
-		f, err := reactApp.Assets.Open("/dist/index.html")
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Fprintf(w, "Error opening React index.html: %v", err)
-			return
-		}
-		defer func() { _ = f.Close() }()
-		idx, err := io.ReadAll(f)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Fprintf(w, "Error reading React index.html: %v", err)
-			return
-		}
-		w.Write(idx)
-	}
-
-	// Static files required by the React app.
-	r.Get("/react-app/*filepath", func(w http.ResponseWriter, r *http.Request) {
-		for _, rt := range reactRouterPaths {
-			if r.URL.Path != "/react-app"+rt {
-				continue
-			}
-			serveReactApp(w, r)
-			return
-		}
-		r.URL.Path = path.Join("/dist", route.Param(r.Context(), "filepath"))
-		fs := server.StaticFileServer(reactApp.Assets)
-		fs.ServeHTTP(w, r)
 	})
 
 	r.Post("/-/reload", func(w http.ResponseWriter, req *http.Request) {
